requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
matplotlib==3.8.2
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
jinja2==3.1.2